+++ Core/Objects/Data/ImageBackgroundObjectData.cs
@@ -1,4 +1,4 @@
-﻿namespace MapImageObjects.Core;
+﻿namespace MapImageObjects.Objects.Data;
 
 // Literally just the same properties as the default image object
 public class ImageBackgroundObjectData : ImageObjectData { }
